Update search activity for Wikipedia app.
The Wikipedia app added an extra search activity in
https://github.com/wikimedia/apps-android-wikipedia/commit/6c188f765b3a6b0fbea86067125d6041c9addf60
and
https://github.com/wikimedia/apps-android-wikipedia/commit/ce9f4447faa94626939073e34a1e5376c33f5816
so change .main.MainActivity to .search.SearchActivity.

Change the search type as well, with 4 it only works if the Wikipedia
app isn't running (second search will just focus the app with whatever
was shown before), with 0 new searches actually search.