Fallback fonts: rework the list management
Add support for storing a default fallback font face list which is
appended when setting a single font face.

The front ends are only setting a single fallback font, but also rely
on a larger fallback font sets which is either the default crengine
list or a platform specific list. This was failing because the code
paths had been setting a single user specified fallback font and then
overriding this with the fallback list.

The code could also get stuck with a filtered list of fallback fonts
and did not retry the filter in the case of new fonts being added. The
new default list is not filtered and when the fallback font changes
they are all retried which now picks up new fonts that might be
available.