Android: workaround fallback font face(s) conflict
The Android reader uses both PROP_FALLBACK_FONT_FACE and
PROP_FALLBACK_FONT_FACES and sets PROP_FALLBACK_FONT_FACES empty if
PROP_FALLBACK_FONT_FACE is defined. This could result in a single face
being set and then the faces being set to an empty set clearing the
initial face set and the fallback font setting not working.

Removing PROP_FALLBACK_FONT_FACES in this case does not solve this
problem as it will be defaulted to a hard coded list and override the
UI fallback font face setting.

The workaround here is to copy the PROP_FALLBACK_FONT_FACE into the
PROP_FALLBACK_FONT_FACES when set.