list marker: separate the content, max width, and rendering functions
renderListItemMarker() had been also computing the maximum width,
walking all the list elements to do so if the value was not cached,
and this part has been split out into a separate function
listItemMarkerMaxWidth() which can simply obtain the cached value if
that is all the caller needs.

getNodeListMarker() had been also computing a width for the marker
content, and that computation has been split out into the few callers,
this leaves this function returning the content of the marker and
brings the code closer to supporting list ::marker content.