android: fix DPI initialization
The DPI was not being initialized correctly leaving it at the default
of 96 dpi, but the intention appears to have been to increase the CSS
DPI for higher DPI devices.

The SettingsManager() was being initialized early, before initializing
the display DPI, and was pushing the initial default DPI of 96 to the
crengine when calling loadSettings(). Move the initialization of the
display DPI earlier to address this.